almost finishing

+++ backend/src/data/mod.rs
@@ -1,7 +1,7 @@
 pub mod models;
 pub mod schema;
 
-pub const CUSTOMER_COUNT: u32 = 30;
+pub const CUSTOMER_COUNT: usize = 30;
 pub const SERVER_COUNT: usize = 4;
